Ported the new systems to fabric

+++ src/main/java/carbonconfiglib/gui/widgets/SuggestionRenderers.java
@@ -117,7 +117,7 @@ public class SuggestionRenderers
 		public Component renderSuggestion(PoseStack stack, String value, int x, int y) {
 			try {
 				GuiComponent.fill(stack, x+1, y+1, x+18, y+19, 0xFFA0A0A0);
-				GuiComponent.fill(stack, x+2, y+2, x+17, y+18, Integer.decode(value) | 0xFF000000);
+				GuiComponent.fill(stack, x+2, y+2, x+17, y+18, Long.decode(value).intValue() | 0xFF000000);
 			}
 			catch(Exception e) {
 			}
